test(payroll): update tests for latest changes
*restructure them under function types
*use mocha contexts from @sablier/dev-utils
