+++ packages/payroll/test/payroll/effects/salary/CreateSalary.js
@@ -20,31 +20,15 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
     describe("when the payroll contract has enough allowance", function() {
       const employee = bob;
       const salary = STANDARD_SALARY.toString(10);
-      let startTime;
-      let stopTime;
-      const isCompounding = false;
-
-      beforeEach(async function() {
-        startTime = now.plus(STANDARD_TIME_OFFSET);
-        stopTime = startTime.plus(STANDARD_TIME_DELTA);
-      });
+      const startTime = now.plus(STANDARD_TIME_OFFSET);
+      const stopTime = startTime.plus(STANDARD_TIME_DELTA);
 
       beforeEach(async function() {
         await this.token.approve(this.payroll.address, salary, opts);
       });
 
       it("creates the salary", async function() {
-        const result = await this.payroll.createSalary(
-          employee,
-          salary,
-          this.token.address,
-          startTime,
-          stopTime,
-          isCompounding,
-          opts,
-        );
-        // We have to force-call the `getSalary` method via the web3.eth.Contract api, otherwise
-        // solidity-coverage will turn it into a state-changing method
+        const result = await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts);
         const salaryObject = await this.payroll.contract.methods.getSalary(Number(result.logs[0].args.salaryId)).call();
         salaryObject.company.should.be.equal(company);
         salaryObject.employee.should.be.equal(employee);
@@ -54,46 +38,34 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
         salaryObject.stopTime.should.be.bignumber.equal(stopTime);
         salaryObject.balance.should.be.bignumber.equal(salary);
         salaryObject.rate.should.be.bignumber.equal(STANDARD_RATE_PER_SECOND);
-        salaryObject.isCompounding.should.be.equal(false);
       });
 
       it("transfers the tokens", async function() {
         const balance = await this.token.balanceOf(company);
-        await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, isCompounding, opts);
+        await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts);
         const newBalance = await this.token.balanceOf(company);
         newBalance.should.be.bignumber.equal(balance.minus(salary));
       });
 
       it("increases the next salary id", async function() {
         const nextSalaryId = await this.payroll.nextSalaryId();
-        await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, isCompounding, opts);
+        await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts);
         const newNextSalaryId = await this.payroll.nextSalaryId();
         newNextSalaryId.should.be.bignumber.equal(nextSalaryId.plus(1));
       });
 
       it("emits an createsalary event", async function() {
-        const result = await this.payroll.createSalary(
-          employee,
-          salary,
-          this.token.address,
-          startTime,
-          stopTime,
-          isCompounding,
-          opts,
-        );
+        const result = await this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts);
         truffleAssert.eventEmitted(result, "CreateSalary");
       });
     });
 
     describe("when the payroll contract does not have enough allowance", function() {
       const employee = bob;
-      let startTime;
-      let stopTime;
-      const isCompounding = false;
+      const startTime = now.plus(STANDARD_TIME_OFFSET);
+      const stopTime = startTime.plus(STANDARD_TIME_DELTA);
 
       beforeEach(async function() {
-        startTime = now.plus(STANDARD_TIME_OFFSET);
-        stopTime = startTime.plus(STANDARD_TIME_DELTA);
         await this.token.approve(this.payroll.address, STANDARD_SALARY.minus(5).toString(10), opts);
       });
 
@@ -102,7 +74,7 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
 
         it("reverts", async function() {
           await truffleAssert.reverts(
-            this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, isCompounding, opts),
+            this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts),
             truffleAssert.ErrorType.REVERT,
           );
         });
@@ -113,7 +85,7 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
 
         it("reverts", async function() {
           await truffleAssert.reverts(
-            this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, isCompounding, opts),
+            this.payroll.createSalary(employee, salary, this.token.address, startTime, stopTime, opts),
             truffleAssert.ErrorType.REVERT,
           );
         });
@@ -124,14 +96,8 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
   describe("when the token contract is not erc20", function() {
     const employee = bob;
     const salary = STANDARD_SALARY.toString(10);
-    let startTime;
-    let stopTime;
-    const isCompounding = false;
-
-    beforeEach(async function() {
-      startTime = now.plus(STANDARD_TIME_OFFSET);
-      stopTime = startTime.plus(STANDARD_TIME_DELTA);
-    });
+    const startTime = now.plus(STANDARD_TIME_OFFSET);
+    const stopTime = startTime.plus(STANDARD_TIME_DELTA);
 
     describe("when the token contract is non-compliant", function() {
       beforeEach(async function() {
@@ -140,15 +106,7 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
 
       it("reverts", async function() {
         await truffleAssert.reverts(
-          this.payroll.createSalary(
-            employee,
-            salary,
-            this.nonStandardERC20Token.address,
-            startTime,
-            stopTime,
-            isCompounding,
-            opts,
-          ),
+          this.payroll.createSalary(employee, salary, this.nonStandardERC20Token.address, startTime, stopTime, opts),
           truffleAssert.ErrorType.REVERT,
         );
       });
@@ -157,7 +115,7 @@ function shouldBehaveLikeCreateSalary(alice, bob) {
     describe("when the token contract is the zero address", function() {
       it("reverts", async function() {
         await truffleAssert.reverts(
-          this.payroll.createSalary(employee, salary, ZERO_ADDRESS, startTime, stopTime, isCompounding, opts),
+          this.payroll.createSalary(employee, salary, ZERO_ADDRESS, startTime, stopTime, opts),
           truffleAssert.ErrorType.REVERT,
         );
       });
